requirements.txt
google-genai
PyMuPDF==1.24.0
openpyxl==3.1.5
Pillow==10.4.0
opencv-python
